v1 merge

+++ src/components/LivePreviewerComponents/SectionItemHeader.tsx
@@ -1,15 +1,17 @@
 import React, { FunctionComponent, useState } from "react";
-import { Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, makeStyles, Typography } from "@material-ui/core";
+import { Box, makeStyles, Typography } from "@material-ui/core";
 import { TooltipIconButton } from "../Material";
 // Icons
 import DeleteIcon from "@material-ui/icons/Delete";
 import EditIcon from "@material-ui/icons/Edit";
+import { colors } from "../../config/theme";
+import { Confirmation } from "../Confirmation/Confirmation";
 
 export interface SectionItemHeaderProps {
   title: string;
   type: string;
   classes?: {
-    actions: string
+    actions: string;
   };
   onDelete: () => void;
   onEdit: () => void;
@@ -31,7 +33,13 @@ export const useSectionItemHeaderStyles = makeStyles({
   },
 });
 
-export const SectionItemHeader: FunctionComponent<SectionItemHeaderProps> = ({ title, type, classes, onDelete, onEdit }) => {
+export const SectionItemHeader: FunctionComponent<SectionItemHeaderProps> = ({
+  title,
+  type,
+  classes,
+  onDelete,
+  onEdit,
+}) => {
   const [deleteConfirmationOpen, setDeleteConfirmationOpen] = useState(false);
 
   return (
@@ -49,39 +57,26 @@ export const SectionItemHeader: FunctionComponent<SectionItemHeaderProps> = ({ t
           tooltip={`Delete ${type}`}
           onClick={() => setDeleteConfirmationOpen(true)}
         >
-          <DeleteIcon fontSize="small" />
+          <DeleteIcon fontSize="small" style={{ color: colors.midBlue }} />
         </TooltipIconButton>
         {/* Edit item */}
-        <TooltipIconButton
-          color="inherit"
-          tooltip={`Edit ${type}`}
-          onClick={onEdit}
-        >
-          <EditIcon fontSize="small" />
+        <TooltipIconButton color="inherit" tooltip={`Edit ${type}`} onClick={onEdit}>
+          <EditIcon fontSize="small" style={{ color: colors.midBlue }} />
         </TooltipIconButton>
       </Box>
 
-      <Dialog
-        open={deleteConfirmationOpen}
-        onClose={() => setDeleteConfirmationOpen(false)}
-      >
-        <DialogTitle>Delete item</DialogTitle>
-        <DialogContent>
-          <Box>Are you sure you want to delete this item?</Box>
-          <Box>This action cannot be reversed.</Box>
-        </DialogContent>
-        <DialogActions>
-          <Button
-            onClick={() => setDeleteConfirmationOpen(false)}
-          >No</Button>
-          <Button
-            onClick={() => {
-              setDeleteConfirmationOpen(false);
-              onDelete();
-            }}
-          >Yes</Button>
-        </DialogActions>
-      </Dialog>
+      <Confirmation
+        isOpen={deleteConfirmationOpen}
+        denyClick={() => setDeleteConfirmationOpen(false)}
+        confirmClick={() => {
+          setDeleteConfirmationOpen(false);
+          onDelete();
+        }}
+        title={"Delete item"}
+        message={`Are you sure you want to delete this item?
+          <br/>
+          This action cannot be reversed.`}
+      />
     </Box>
   );
 };
